changed the data type of the angle from Double to Int

+++ src/simulation/2DTissue.h
@@ -65,10 +65,10 @@ private:
     Eigen::MatrixXd r_3D;
     Eigen::MatrixXd r_3D_old;
     Eigen::Matrix<double, Eigen::Dynamic, 2> r_dot;
-    Eigen::VectorXd n;
+    Eigen::VectorXi n;
     Eigen::Vector2d original_pole;
-    Eigen::VectorXd n_pole;
-    Eigen::VectorXd n_pole_old;
+    Eigen::VectorXi n_pole;
+    Eigen::VectorXi n_pole_old;
     std::vector<int> particles_color;
     std::vector<int> vertices_3D_active;
     Eigen::MatrixXd distance_matrix;
