+++ src/simulation/VirtualMesh.h
@@ -31,7 +31,7 @@ public:
         Eigen::MatrixXd& vertice_UV,
         std::vector<int64_t>& h_v_mapping,
         int particle_count,
-        Eigen::VectorXd& n,
+        Eigen::VectorXi& n,
         Eigen::MatrixXi& face_3D,
         Eigen::MatrixXd& vertice_3D,
         Eigen::MatrixXd& distance_matrix,
@@ -43,9 +43,9 @@ public:
 
     void prepare_virtual_mesh(int old_id);
     Eigen::Vector2d init_north_pole();
-    Eigen::VectorXd get_relative_orientation();
-    void assign_particle_orientation(Eigen::Vector2d northPole_, Eigen::VectorXd n_pole_);
-    Eigen::VectorXd get_n_orientation(Eigen::Matrix<double, Eigen::Dynamic, 2> position_, Eigen::Vector2d northPole_, Eigen::VectorXd n_pole_);
+    Eigen::VectorXi get_relative_orientation();
+    void assign_particle_orientation(Eigen::Vector2d northPole_, Eigen::VectorXi n_pole_);
+    Eigen::VectorXi get_n_orientation(Eigen::Matrix<double, Eigen::Dynamic, 2> position_, Eigen::Vector2d northPole_, Eigen::VectorXi n_pole_);
     void change_UV_map(int target_vertex);
 
 private:
@@ -57,7 +57,7 @@ private:
     Eigen::MatrixXd& vertice_UV;
     std::vector<int64_t>& h_v_mapping;
     int particle_count;
-    Eigen::VectorXd& n;
+    Eigen::VectorXi& n;
     Eigen::MatrixXi& face_3D;
     Eigen::MatrixXd& vertice_3D;
     Eigen::MatrixXd& distance_matrix;
